+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/edit/AdvancedSettings.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LogicEditor from "@/app/environments/[environmentId]/surveys/[surveyId]/edit/LogicEditor";
+import LogicEditor from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/edit/LogicEditor";
 import UpdateQuestionId from "./UpdateQuestionId";
 import { Question } from "@formbricks/types/questions";
 import { Survey } from "@formbricks/types/surveys";
