--- apps/web/app/(app)/invite/layout.tsx
@@ -1,5 +1,0 @@
-import { PosthogClientWrapper } from "../PosthogClientWrapper";
-
-export default async function AuthLayout({ children }: { children: React.ReactNode }) {
-  return <PosthogClientWrapper>{children}</PosthogClientWrapper>;
-}
