+++ apps/web/app/(app)/layout.tsx
@@ -1,13 +1,29 @@
-import { PHProvider, PostHogPageview } from "./PostHogClient";
+import FormbricksClient from "@/app/(app)/FormbricksClient";
+import { PHProvider, PostHogPageview } from "@/app/PostHogClient";
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
 import { Suspense } from "react";
+import PosthogIdentify from "./PosthogIdentify";
 
-export default function AppLayout({ children }) {
+export default async function AppLayout({ children }) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return redirect(`/auth/login`);
+  }
+
   return (
     <>
       <Suspense>
         <PostHogPageview />
       </Suspense>
-      <PHProvider>{children}</PHProvider>
+      <PHProvider>
+        <>
+          <PosthogIdentify session={session} />
+          <FormbricksClient session={session} />
+          {children}
+        </>
+      </PHProvider>
     </>
   );
 }
