+++ apps/web/app/(app)/onboarding/page.tsx
@@ -1,14 +1,8 @@
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 import { getServerSession } from "next-auth";
 import Onboarding from "./Onboarding";
-import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
-import FormbricksClient from "@/app/FormbricksClient";
 
 export default async function OnboardingPage() {
   const session = await getServerSession(authOptions);
-  return (
-    <>
-      <FormbricksClient session={session} />
-      <Onboarding session={session} />
-    </>
-  );
+  return <Onboarding session={session} />;
 }
