+++ apps/web/app/(auth)/auth/forgot-password/email-sent/page.tsx
