+++ apps/web/app/(auth)/auth/forgot-password/page.tsx
@@ -1,4 +1,4 @@
-import { PasswordResetForm } from "../../../components/auth/RequestPasswordResetForm";
+import { PasswordResetForm } from "@/components/auth/RequestPasswordResetForm";
 import FormWrapper from "@/components/auth/FormWrapper";
 
 const ForgotPasswordPage: React.FC = () => {
