+++ apps/web/app/(auth)/auth/forgot-password/reset/page.tsx
