+++ apps/web/app/(auth)/layout.tsx
@@ -1,0 +1,13 @@
+import { PHProvider, PostHogPageview } from "../PostHogClient";
+import { Suspense } from "react";
+
+export default function AppLayout({ children }) {
+  return (
+    <>
+      <Suspense>
+        <PostHogPageview />
+      </Suspense>
+      <PHProvider>{children}</PHProvider>
+    </>
+  );
+}
