+++ apps/web/components/team/CreateTeamModal.tsx
@@ -8,7 +8,7 @@ import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
-import { createTeam } from "../../app/environments/[environmentId]/actions";
+import { createTeam } from "@/app/(app)/environments/[environmentId]/actions";
 
 interface CreateTeamModalProps {
   open: boolean;
