+++ apps/web/lib/surveys/surveys.ts
@@ -5,11 +5,14 @@ import { TResponse } from "@formbricks/types/v1/responses";
 import {
   OptionsType,
   QuestionOptions,
-} from "@/app/environments/[environmentId]/surveys/[surveyId]/QuestionsComboBox";
-import { QuestionFilterOptions } from "@/app/environments/[environmentId]/surveys/[surveyId]/ResponseFilter";
+} from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/QuestionsComboBox";
+import { QuestionFilterOptions } from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/ResponseFilter";
 import { QuestionType } from "@formbricks/types/questions";
 import { TTag } from "@formbricks/types/v1/tags";
-import { DateRange, SelectedFilterValue } from "@/app/environments/[environmentId]/ResponseFilterContext";
+import {
+  DateRange,
+  SelectedFilterValue,
+} from "@/app/(app)/environments/[environmentId]/ResponseFilterContext";
 import { isArray } from "lodash";
 import { isWithinInterval } from "date-fns";
